LayoutTests/http/tests/security/contentSecurityPolicy/xsl-img-blocked-expected.txt

CONSOLE MESSAGE: Refused to load image from 'http://127.0.0.1:8000/security/resources/abe.png' because of Content-Security-Policy.

Here is an image:  
